Extract record loading controls.
