+++ app-modules/assist-data-model/src/Filament/Resources/StudentResource/Pages/StudentEngagementTimeline.php
@@ -6,7 +6,6 @@ use Assist\Engagement\Models\Engagement;
 use Assist\Timeline\Filament\Pages\Timeline;
 use Assist\Timeline\Actions\SyncTimelineData;
 use Assist\Engagement\Models\EngagementResponse;
-use Assist\Timeline\Models\Timeline as TimelineModel;
 use Assist\AssistDataModel\Filament\Resources\StudentResource;
 
 class StudentEngagementTimeline extends Timeline
@@ -33,21 +32,5 @@ class StudentEngagementTimeline extends Timeline
         // 1. Utilizing the updated at of the timeline record
         // 2. Grabbing the last timelineable model and checking if it's in the timeline
         resolve(SyncTimelineData::class)->now($this->recordModel, $this->modelsToTimeline);
-    }
-
-    protected function getViewData(): array
-    {
-        $timelineRecords = TimelineModel::query()
-            ->forEducatable($this->recordModel)
-            ->whereIn(
-                'timelineable_type',
-                collect($this->modelsToTimeline)->map(fn ($model) => resolve($model)->getMorphClass())->toArray()
-            )
-            ->orderBy('record_creation', 'desc')
-            ->simplePaginate($this->recordsPerPage);
-
-        return [
-            'timelineRecords' => $timelineRecords,
-        ];
     }
 }
