+++ app-modules/timeline/src/Filament/Pages/Concerns/LoadsRecords.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Assist\Timeline\Filament\Pages\Concerns;
+
+trait LoadsRecords
+{
+    public int $recordsPerPage = 3;
+
+    public function loadMoreRecords(): void
+    {
+        $this->recordsPerPage += 3;
+    }
+}
